pyproject.toml
[project]
name = "midscene-agent"
version = "1.0.0"
description = "AI-driven web automation framework"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "AI Automation Team", email = "team@example.com" }]
keywords = ["automation", "web", "ai", "langgraph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-deepseek>=1.0.0",
    "aiohttp>=3.9.0",
    "asyncio-throttle>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
    "langgraph-cli[inmem]>=0.4.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
midscene = "runner.run:main_wrapper"
midscene-check = "runner.check_config:check_config"
midscene-yaml = "runner.executor.yaml_executor:main"
midscene-text = "runner.executor.text_executor:main"
midscene-format = "runner.format:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["runner"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
